RCL-369 Delete a recall (#200)
* Adds recall deletion functionality

Enables users to delete recalls from the system.

Introduces a confirmation page to prevent accidental deletions.
Only the latest recall is eligible for deletion, as indicated by its creation date.
When a recall is deleted, any associated UAL information is also removed.

* Displays prisoner details on delete confirmation

* Create tests for viewPersonHome.ts

* Create tests for viewPersonHome.ts

* Adds delete recall confirmation tests

Adds tests for the delete recall confirmation functionality,
covering both GET and POST requests to ensure the
confirmation page is rendered correctly and that recalls
can be successfully deleted or the deletion process can be cancelled.

* removed empty line

* Adds eslint disable comment

Adds an eslint disable comment to prevent import errors on CI.

* Adds eslint disable for unresolved imports

Adds an eslint disable comment for unresolved imports
in controller test and non-test files, to suppress
the "import/no-unresolved" error.