RCLL-338 excludes current recall from validation (#205)
* Excludes current recall's UAL from validation

When editing a recall, this change ensures that the validation logic excludes UAL (Unlawfully at Large) adjustments associated with the current recall being edited. This prevents conflicts when determining valid revocation or return to custody dates, as the existing UAL for the current recall should not be considered as a conflict. It also fixes a bug where the revocation date could be set before the earliest sentence date.

* Refactors UAL adjustment validation logic

Centralizes the logic for determining which adjustments to consider during UAL validation into a single function, `getAdjustmentsToConsiderForValidation`.

This improves code reuse and consistency across revocation and return to custody date controllers, reducing redundancy and improving maintainability.

* Updates test descriptions for clarity

Refines test descriptions in recall revocation and return to custody date controller tests for better clarity and readability.

Also removes an unused import from returnToCustodyDateController.test.ts.

* Removes unused import

Removes the unused import of the eslint-disable-next-line comment.

* Updates mock for form wizard helper