Adds Manage Offences API integration (#217)
* Adds Manage Offences API integration

Introduces the Manage Offences API client and service.

This change allows the application to interact with the Manage Offences API, providing functionalities such as retrieving offence details and searching for offences.

* Adds manage offences service to person home

* Adds Manage Offences API URL

Adds the Manage Offences API URL to the list of API URLs used by the application in dev, preprod, and prod environments.